Fixes to timestamp support.
Factor the timestamp parsing into its own function that
can easily be changed later to support the full range of
time stamps and add a quick early check to avoid
going through all the time.Parse calls in this common
path.

Also fix marshaling so that times are still printed unquoted.

For backwards compatibility, when timestamps are unmarshaled
into interface{} values, we unmarshal them as strings.

We also make it possible to unmarshal any value that implements
TextUnmarshaler, regardless of any explicit or implicit tag.